Int32 batched joint counts + fused int32 entropy reduce (the 34%-of-GPU kernel)
nsys at the full F2 1M fit put rows_ent_nnz at 9.35s -- 34% of ALL GPU time -- reading float64 cells of the
huge, mostly-zero (K, Kx*Kyz) CMI counts matrix, after the callers' int64->float64 .astype had already
copied the same matrix once more. The counts are exact cell frequencies <= n < 2^31.

* batched_joint_hist1/2 gain int32 twins (plain int atomics); _batched_joint_counts2 /
  _batched_marginal_counts allocate int32 -- half the memset, half the buffer VRAM (the perm-null's dense
  joint fits 2x the perms per chunk; its VRAM sizer updated to *4);
* rows_ent_nnz gains an int32 variant casting in-register -- the .astype(float64) copies at the
  batched_cmi_gpu call sites are gone and the reduce reads 4 B/cell instead of 8;
* identical counts -> identical entropies/cardinalities -> bit-identical CMI.

Measured (nsys, full 1M fit): the entropy reduce 9.35s/242 launches -> 5.89s/162 (wider chunks), the
batched hist halves, ~3.5s less GPU time on the CMI path. F2 across distributions/scale/single-step +
raw-redundancy suites: 19 passed, 1 xfailed (pre-existing), same compound; wall within the thermal band
(the saved GPU time overlaps host phases on this card).

+++ src/mlframe/feature_selection/filters/_fe_batched_mi.py
@@ -529,9 +529,9 @@ def batched_cmi_gpu(x_cols: Any, y: np.ndarray, z: Any = None, return_cards: boo
         # Marginal MI(x_k; y), MM-corrected:  H(x)+H(y)-H(x,y) - (k_x+k_y-k_xy-1)/2n
         Ky = int(dy.max()) + 1 if dy.size else 1
         _assert_codes_in_range(dy, Ky, "batched_cmi_gpu y codes", codes_trusted)
-        cnt_xy = _batched_joint_counts2(X, dy, Kx, Ky).astype(cp.float64)   # (K, Kx*Ky), in-kernel flat key
+        cnt_xy = _batched_joint_counts2(X, dy, Kx, Ky)   # (K, Kx*Ky) int32; the reduce casts in-register
         h_xy, k_xy = _rows_entropy_and_k(cnt_xy, inv_n)
-        cnt_x = _batched_marginal_counts(X, Kx).astype(cp.float64)          # (K, Kx)
+        cnt_x = _batched_marginal_counts(X, Kx)          # (K, Kx) int32
         h_x, k_x = _rows_entropy_and_k(cnt_x, inv_n)
         h_y, k_y = joint_entropy_gpu([dy], [Ky], inv_n)                     # H(y): fused hist+entropy in ONE launch (same 2-launch fallback if it won't fit shared)
         # FUSED assembly: max(h_x - h_xy + h_y - (k_x - k_xy + (k_y-1))/2n, 0) in one (K,) launch.
@@ -574,14 +574,14 @@ def batched_cmi_gpu(x_cols: Any, y: np.ndarray, z: Any = None, return_cards: boo
     if _fused is not None:
         h_xz, k_xz = _fused
     else:
-        cnt_xz = _batched_joint_counts2(X, dz, Kx, Kz).astype(cp.float64)   # (K, Kx*Kz)
+        cnt_xz = _batched_joint_counts2(X, dz, Kx, Kz)   # (K, Kx*Kz) int32
         h_xz, k_xz = _rows_entropy_and_k(cnt_xz, inv_n)
     # H(x_k, y, z) for all k: in-kernel flat key k*(Kx*Kyz) + x*Kyz + yz (same fuse; large Kyz -> fallback)
     _fused = _batched_joint_entropy_and_k2(X, yz, Kx, Kyz, inv_n)
     if _fused is not None:
         h_xyz, k_xyz = _fused
     else:
-        cnt_xyz = _batched_joint_counts2(X, yz, Kx, Kyz).astype(cp.float64) # (K, Kx*Kyz)
+        cnt_xyz = _batched_joint_counts2(X, yz, Kx, Kyz) # (K, Kx*Kyz) int32
         h_xyz, k_xyz = _rows_entropy_and_k(cnt_xyz, inv_n)
 
     # FUSED assembly: max(h_xz - h_xyz + (h_yz - h_z) - (k_xyz - k_xz + (k_z - k_yz))/2n, 0) in one (K,) launch.
